+++ Changelog.md
@@ -2,6 +2,11 @@
 
 The changes for each individual (beta) release can be seen [here](https://github.com/christian-photo/ninaAPI/releases). However this Changelog will be mostly complete.
 
+## 2.2.2.0
+
+- Added `ROTATOR-MOVED` and `ROTATOR-MOVED-MECHANICAL` events to the websocket
+- Added `application/get-tab` to get the current application tab
+
 ## 2.2.1.0
 
 ### Sequence